+++ src/components/Molecules/Header/Header.style.js
@@ -1,7 +1,10 @@
+import React, { useState } from 'react';
 import styled from 'styled-components';
 
 import Link from '../../Atoms/Link/Link';
+import Text from '../../Atoms/Text/Text';
 import hideVisually from '../../../theme/shared/hideVisually';
+import { BurgerIcon, BurgerMenu as Wrapper } from './BurgerMenu.style';
 
 const HeaderWrapper = styled.header.attrs(() => ({
   role: 'Banner'
@@ -207,7 +210,33 @@ const MetaIcons = styled.div`
   }
 `;
 
+const BurgerMenu = () => {
+  const [toggleState, setToggleState] = useState(false);
+  function toggle() {
+    setToggleState(!toggleState);
+  }
+  return (
+    <Wrapper>
+      <NavLink
+        href="#"
+        role="button"
+        aria-expanded={toggleState}
+        aria-haspopup="true"
+        aria-label="Open and close Navigation Menu, 24 items listed"
+        inline
+        onClick={toggle}
+      >
+        <BurgerIcon>
+          <Text>Open and close nav menu</Text>
+        </BurgerIcon>
+        <span />
+      </NavLink>
+    </Wrapper>
+  );
+};
+
 export {
+  BurgerMenu,
   Brand,
   HeaderWrapper,
   InnerWrapper,
